+++ src/applications/disability-benefits/996/components/GetFormHelp.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 const GetFormHelp = () => (
   <p className="help-talk">
-    Need help filling out the form or have questions about eligibility? please
+    Need help filling out the form or have questions about eligibility? Please
     call VA Benefits and Services at{' '}
     <a className="nowrap" href="tel:1-800-827-1000">
       800-827-1000
